refactor(logs): lift PostHogLogs to core and isolate storage failures
- Move PostHogLogs class + types from packages/react-native to packages/core
  so js-lite/node can reuse the same orchestrator when they add logs (matches
  the ErrorTracking pattern: shared primitives in core, per-SDK class where
  needed).
- Make PostHogCoreStateless.wrap() public (with @internal JSDoc) so the
  composition class can reuse core init-gating instead of mirroring the
  _initPromise.then(fn) pattern locally.
- Switch storage preloads from Promise.all to Promise.allSettled with
  per-side error logging. Events-storage failure no longer blocks logs (and
  vice versa); a failing side degrades to memory-only via PostHogRNStorage
  persist() error handling.
- Split config: interface (PostHogLogsConfig + ResolvedPostHogLogsConfig) in
  core; default constants in packages/react-native/src/logs-defaults.ts. Each
  SDK owns its own tuning (mobile cellular costs vs browser vs node).
- Move optedOut fast-exit to top of captureLog (saves OTLP record build when
  already opted out), keep inner re-check in _enqueue for pre-init-stale
  correctness.
- Drop PostHogLogs.flushStorage(); RN AppState handler now calls
  _logsStorage.waitForPersist() directly (RN-specific shutdown concern, not
  part of the core class).

Tests: core 468/468 pass (22 logs tests moved over); RN 282/282 pass.
