+++ packages/core/src/posthog-core-stateless.ts
@@ -194,7 +194,15 @@ export abstract class PostHogCoreStateless {
     }
   }
 
-  protected wrap(fn: () => void): void {
+  /**
+   * Schedule fn to run once the SDK has finished initializing. If init has
+   * been rejected or the SDK is disabled, fn is silently dropped. Used by
+   * the SDK's own capture/identify/alias/etc. paths, and by composition
+   * modules (e.g. PostHogLogs) that need init-gated work without subclassing.
+   *
+   * @internal
+   */
+  public wrap(fn: () => void): void {
     if (this.disabled) {
       this._logger.warn('The client is disabled')
       return
